Updated the unrooted option, should now use a pure towelroot to install on devices which do not have sony_ric kernel protection
Updated PhilZ, various times.

Updated TWRP for some devices

Updated CWM for some devices

+++ src/install.bat
@@ -165,43 +165,62 @@ echo Device model is %product_name%
 for /f "delims=" %%i in ('adb shell "getprop ro.build.id"') do ( set firmware=%%i)
 echo Firmware is %firmware%
 
-echo.
-echo =============================================
-echo Sending files
-echo =============================================
+for /f "delims=" %%i in ('adb shell "/data/local/tmp/recovery/busybox grep '/sys/kernel/security/sony_ric/enable' init.* | /data/local/tmp/recovery/busybox wc -l") do ( set sony_ric=%%i )
 
-set zxzFile=zxz.sh
+if %sony_ric% = 1 (
 
-for /f "delims=" %%i in ('adb shell "ls /dev/kmem"') do ( set kmem_exist=%%i )
-if NOT "%kmem_exist%" == "/dev/kmem: No such file or directory " (
-  set zxzFile="zxz_kmem.sh"
-  echo /dev/kmem exists. Using files of cubeundcube...
+	echo.
+	echo =============================================
+	echo Sending files
+	echo =============================================
+
+	set zxzFile=zxz.sh
+
+	for /f "delims=" %%i in ('adb shell "ls /dev/kmem"') do ( set kmem_exist=%%i )
+	if NOT "%kmem_exist%" == "/dev/kmem: No such file or directory " (
+	  set zxzFile="zxz_kmem.sh"
+	  echo /dev/kmem exists. Using files of cubeundcube...
+	)
+
+	adb push easyroottool\%zxzFile% /data/local/tmp/zxz.sh
+	adb push easyroottool\writekmem /data/local/tmp/
+	adb push easyroottool\findricaddr /data/local/tmp/
+	adb shell "cp /data/local/tmp/recovery/busybox /data/local/tmp/"
+	adb shell "chmod 777 /data/local/tmp/busybox"
+	adb shell "chmod 777 /data/local/tmp/zxz.sh"
+	adb shell "chmod 777 /data/local/tmp/writekmem"
+	adb shell "chmod 777 /data/local/tmp/findricaddr"
+
+	if "%zxzFile%" == ""zxz.sh"" (
+	  echo.
+	  echo Copying kernel module...
+	  adb push easyroottool\kernelmodule_patch.sh /data/local/tmp/kernelmodule_patch.sh
+	  adb shell "chmod 777 /data/local/tmp/kernelmodule_patch.sh"
+	  adb shell "/data/local/tmp/kernelmodule_patch.sh"
+	)
+
+	set towelrootbinary=easyroottool/tr_signed.apk
+
+	echo.
+	echo =============================================
+	echo Loading geohot's towelroot ^(modified by zxz0O0^)
+	echo =============================================
+
 )
 
-adb push easyroottool\%zxzFile% /data/local/tmp/zxz.sh
-adb push easyroottool\writekmem /data/local/tmp/
-adb push easyroottool\findricaddr /data/local/tmp/
-adb shell "cp /data/local/tmp/recovery/busybox /data/local/tmp/"
-adb shell "chmod 777 /data/local/tmp/busybox"
-adb shell "chmod 777 /data/local/tmp/zxz.sh"
-adb shell "chmod 777 /data/local/tmp/writekmem"
-adb shell "chmod 777 /data/local/tmp/findricaddr"
+if %sony_ric% = 0 (
 
-if "%zxzFile%" == ""zxz.sh"" (
-  echo.
-  echo Copying kernel module...
-  adb push easyroottool\kernelmodule_patch.sh /data/local/tmp/kernelmodule_patch.sh
-  adb shell "chmod 777 /data/local/tmp/kernelmodule_patch.sh"
-  adb shell "/data/local/tmp/kernelmodule_patch.sh"
+	set towelrootbinary=easyroottool/tr.apk
+
+	echo.
+	echo =============================================
+	echo Loading geohot's towelroot
+	echo =============================================
+
 )
-
-echo.
-echo =============================================
-echo Loading geohot's towelroot ^(modified by zxz0O0^)
-echo =============================================
 
 adb uninstall com.geohot.towelroot
-adb install easyroottool\tr_signed.apk
+adb install %towelrootbinary%
 
 adb shell "am start -n com.geohot.towelroot/.TowelRoot" >nul 2>&1
 echo =============================================
@@ -219,8 +238,7 @@ if "%isRooted%" == "/system/bin/sh: su: not found" goto RootCheck
 if "%isRooted%" == """" goto RootCheck
 echo.
 adb uninstall com.geohot.towelroot
-adb shell "su -c /data/local/tmp/recovery/install.sh"
-adb shell "su -c rm /system/etc/install-recovery.sh /system/xbin/*su"
+adb shell "su -c /data/local/tmp/recovery/install.sh unrooted"
 goto cleanup
 
 :end
